# docs/openapi.yaml
{
  "components": {
    "schemas": {
      "CreateSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "ttl_seconds": {
            "default": 600,
            "maximum": 1800.0,
            "minimum": 300.0,
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "title": "CreateSessionRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "RenewSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "ttl_seconds": {
            "default": 600,
            "maximum": 1800.0,
            "minimum": 300.0,
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "RenewSessionRequest",
        "type": "object"
      },
      "RevokeSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "reason"
        ],
        "title": "RevokeSessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "credential_delivered": {
            "default": true,
            "title": "Credential Delivered",
            "type": "boolean"
          },
          "credential_rotated": {
            "default": false,
            "title": "Credential Rotated",
            "type": "boolean"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "mock_mode": {
            "const": true,
            "default": true,
            "title": "Mock Mode",
            "type": "boolean"
          },
          "realm": {
            "title": "Realm",
            "type": "string"
          },
          "renew_after": {
            "format": "date-time",
            "title": "Renew After",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "sip_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sip Password"
          },
          "sip_username": {
            "title": "Sip Username",
            "type": "string"
          },
          "storage_requirement": {
            "const": "memory-only",
            "default": "memory-only",
            "title": "Storage Requirement",
            "type": "string"
          },
          "wss_url": {
            "title": "Wss Url",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "endpoint",
          "sip_username",
          "sip_password",
          "wss_url",
          "realm",
          "expires_at",
          "renew_after",
          "request_id",
          "correlation_id"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Trusted JWT/OIDC access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Codestra SIP Provisioning API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/sip/config": {
      "get": {
        "operationId": "sip_config_api_v1_sip_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Sip Config Api V1 Sip Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Sip Config"
      }
    },
    "/api/v1/sip/session": {
      "delete": {
        "operationId": "revoke_api_v1_sip_session_delete",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Client-Instance-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Client-Instance-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke"
      },
      "post": {
        "operationId": "create_api_v1_sip_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Client-Instance-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Client-Instance-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create"
      }
    },
    "/api/v1/sip/session/renew": {
      "post": {
        "operationId": "renew_api_v1_sip_session_renew_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Client-Instance-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Client-Instance-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Renew"
      }
    },
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    },
    "/readyz": {
      "get": {
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readyz"
      }
    }
  }
}